✨ transform note content to markdown format
